+++ portal-ejb/src/com/liferay/portal/service/persistence/OrganizationFinder.java
@@ -105,7 +105,7 @@ public class OrganizationFinder {
 
 			SQLQuery q = session.createSQLQuery(sql);
 
-			q.addScalar(HibernateUtil.getCountColumnName(), Hibernate.INTEGER);
+			q.addScalar(HibernateUtil.getCountColumnName(), Hibernate.LONG);
 
 			QueryPos qPos = QueryPos.getInstance(q);
 
@@ -137,7 +137,7 @@ public class OrganizationFinder {
 			Iterator itr = q.list().iterator();
 
 			if (itr.hasNext()) {
-				Integer count = (Integer)itr.next();
+				Long count = (Long)itr.next();
 
 				if (count != null) {
 					return count.intValue();
@@ -287,7 +287,7 @@ public class OrganizationFinder {
 
 			SQLQuery q = session.createSQLQuery(sql);
 
-			q.addScalar(HibernateUtil.getCountColumnName(), Hibernate.INTEGER);
+			q.addScalar(HibernateUtil.getCountColumnName(), Hibernate.LONG);
 
 			QueryPos qPos = QueryPos.getInstance(q);
 
@@ -329,7 +329,7 @@ public class OrganizationFinder {
 			Iterator itr = q.list().iterator();
 
 			while (itr.hasNext()) {
-				Integer i = (Integer)itr.next();
+				Long i = (Long)itr.next();
 
 				if (i != null) {
 					count += i.intValue();
