+++ portal-ejb/src/com/liferay/portal/service/persistence/PermissionUserFinder.java
@@ -118,7 +118,7 @@ public class PermissionUserFinder {
 
 			SQLQuery q = session.createSQLQuery(sql);
 
-			q.addScalar(HibernateUtil.getCountColumnName(), Hibernate.INTEGER);
+			q.addScalar(HibernateUtil.getCountColumnName(), Hibernate.LONG);
 
 			QueryPos qPos = QueryPos.getInstance(q);
 
@@ -130,7 +130,7 @@ public class PermissionUserFinder {
 			Iterator itr = q.list().iterator();
 
 			if (itr.hasNext()) {
-				Integer count = (Integer)itr.next();
+				Long count = (Long)itr.next();
 
 				if (count != null) {
 					return count.intValue();
@@ -252,7 +252,7 @@ public class PermissionUserFinder {
 
 		SQLQuery q = session.createSQLQuery(sql);
 
-		q.addScalar(HibernateUtil.getCountColumnName(), Hibernate.INTEGER);
+		q.addScalar(HibernateUtil.getCountColumnName(), Hibernate.LONG);
 
 		QueryPos qPos = QueryPos.getInstance(q);
 
@@ -284,7 +284,7 @@ public class PermissionUserFinder {
 		Iterator itr = q.list().iterator();
 
 		if (itr.hasNext()) {
-			Integer count = (Integer)itr.next();
+			Long count = (Long)itr.next();
 
 			if (count != null) {
 				return count.intValue();
