+++ portal-ejb/src/com/liferay/portlet/documentlibrary/service/persistence/DLFileEntryPersistence.java
@@ -439,7 +439,7 @@ public class DLFileEntryPersistence extends BasePersistence {
 			Iterator itr = q.list().iterator();
 
 			if (itr.hasNext()) {
-				Integer count = (Integer)itr.next();
+				Long count = (Long)itr.next();
 
 				if (count != null) {
 					return count.intValue();
